requirements.txt
astroid==2.4.2
certifi==2020.11.8
chardet==3.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
prompt-toolkit==3.0.8
pyfiglet==0.8.post1
Pygments==2.7.3
pylint==2.6.0
python-dotenv==0.15.0
python-vlc==3.0.11115
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
wcwidth==0.2.5
wrapt==1.12.1
yapf==0.30.0
